Add entry-move signal to media-store
Create a new signal for media stores that allow
it to signal the system to change the media type
of the entry and put it in the new store.  This is
helpful for things like videos where they are initially
imported as movies then have to be moved to the actual
type.